Struct hash_roll::Zpaq

pub struct Zpaq {
    // some fields omitted
}

zpaq - a splitter used in go 'dedup' and zpaq that does not require looking back in the source data to update

PDF: ??

Note: go-dedup & zpaq calculate the relationship between their parameters slightly differently. We support both of these (via the seperate with_*() constructors, but it'd be nice to clarify why they differ and what affect the differences have.

References: http://encode.ru/threads/456-zpaq-updates?p=45192&viewfull=1#post45192 https://github.com/klauspost/dedup/blob/master/writer.go#L668 'zpaqWriter' https://github.com/zpaq/zpaq/blob/master/zpaq.cpp

Parameters: fragment (aka average_size_base_2): average size = 2**fragment KiB in Zpaq (the compressor), this defaults to 6 min_size, max_size: additional bounds on the blocks. Not technically needed for the algorithm to function

In Zpaq-compressor, min & max size are calculated using the fragment value In go's dedup, fragment is calculated using a min & max size

In-block state: hash: u32, current hash last_byte: u8, previous byte read predicted_byte: array of 256 u8's.

Between-block state: None

Methods

impl Zpaq

fn with_range(range: Range<usize>) -> Self

Create a splitter using the range of output block sizes.

The average block size will be the max block size (if any) divided by 4, using the same algorithm to calculate it as go-dedup.

fn with_average_size(average_size_base_2: u8) -> Self

Create a splitter using the defaults from Zpaq (the compressor) given a average size is base 2 (zpaq argument "-fragment")

fn with_max_size(max: usize) -> Self

Use the defaults from go-dedup to generate a splitter given the max size of a split.

The average block size will be the max block size (if any) divided by 4, using the same algorithm to calculate it as go-dedup.

fn with_average_and_range(average_size_base_2: u8, range: Range<usize>) -> Self

Create a splitter with control of all parameters

All the other constructors use this internally

fn new() -> Self

Create a splitter using the defaults from Zpaq (the compressor)

Average size is 65536 bytes (64KiB), max is 520192 bytes (508KiB), min is 4096 bytes (4KiB)