Trait hash_roll::Splitter

pub trait Splitter {
    fn split<'b>(&self, data: &'b [u8]) -> (&'b [u8], &'b [u8]);
    fn next_iter<T: Iterator<Item=u8>>(&self, iter: T) -> Option<Vec<u8>>;

    fn into_slices<'a>(self, data: &'a [u8]) -> SplitterSplit<'a, Self> where Self: Sized { ... }
    fn as_slices<'a>(&'a self, data: &'a [u8]) -> SplitterSplit<'a, &Self> where Self: Sized { ... }
    fn into_vecs<'a, T: Iterator<Item=u8>>(self, data: T) -> SplitterVecs<T, Self> where Self: Sized { ... }
    fn as_vecs<'a, T: Iterator<Item=u8>>(&'a self, data: T) -> SplitterVecs<T, &Self> where Self: Sized { ... }
}

Required Methods

fn split<'b>(&self, data: &'b [u8]) -> (&'b [u8], &'b [u8])

Split data into 2 pieces using a given splitter.

It is expected that in most cases the second element of the return value will be split further by calling this function again.

fn next_iter<T: Iterator<Item=u8>>(&self, iter: T) -> Option<Vec<u8>>

Return chunks from a given iterator, split according to the splitter used.

Provided Methods

fn into_slices<'a>(self, data: &'a [u8]) -> SplitterSplit<'a, Self> where Self: Sized

fn as_slices<'a>(&'a self, data: &'a [u8]) -> SplitterSplit<'a, &Self> where Self: Sized

fn into_vecs<'a, T: Iterator<Item=u8>>(self, data: T) -> SplitterVecs<T, Self> where Self: Sized

fn as_vecs<'a, T: Iterator<Item=u8>>(&'a self, data: T) -> SplitterVecs<T, &Self> where Self: Sized